gradient now watched

+++ PyraPose/initializers.py
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 """
 
-import keras
+import tensorflow.keras as keras
 
 import numpy as np
 import math
@@ -34,7 +34,7 @@ class PriorProbability(keras.initializers.Initializer):
 
     def __call__(self, shape, dtype=None):
         # set bias to -log((1 - p)/p) for foreground
-        result = np.ones(shape, dtype=dtype) * -math.log((1 - self.probability) / self.probability)
-        #result = np.ones(shape, dtype=np.float32) * -math.log((1 - self.probability) / self.probability)
+        #result = np.ones(shape, dtype=dtype) * -math.log((1 - self.probability) / self.probability)
+        result = np.ones(shape, dtype=np.float32) * -math.log((1 - self.probability) / self.probability)
 
         return result
